Added 'hub' submodule for reusable contracts; split 256-game from generic Bisection contract
